Split up controller into seperate controllers
